+++ packages/frontend/src/shared/httpClient/HttpProvider.ts
@@ -1,4 +1,5 @@
 import axios from 'axios';
+import { useToast } from 'vue-toastification';
 
 const serverPort = (import.meta.env.SERVER_PORT || '3000') as string;
 const serverAddress = (import.meta.env.SERVER_ADDRESS ||
@@ -6,10 +7,11 @@ const serverAddress = (import.meta.env.SERVER_ADDRESS ||
 const serverUrl = `${window.location.protocol}//${serverAddress}:${serverPort}/api/`;
 
 export const httpClient = axios.create({
+  withCredentials: true,
   baseURL: serverUrl,
   timeout: 1000,
   validateStatus: function (status) {
-    return status >= 200 && status < 300; // default
+    return (status >= 200 && status < 300) || status == 401; // default
   },
 });
 
@@ -20,7 +22,8 @@ httpClient.interceptors.response.use(
 
   function onRejected(error) {
     console.trace(error);
-    alert('Načítání dat selhalo.');
+    const toast = useToast();
+    toast.error('Server error.');
     return Promise.reject(new Error(JSON.stringify(error)));
   },
 );