+++ docs/changelog.md
@@ -1,4 +1,14 @@
-v0.1.0
+# v0.2.0
+Version introduces authentication and user accounts and also a 
+new features for Bible reader
+
+## Features:
+- Authentication with email and password
+- Verse highlighting
+- Verse notes
+- Bookmarking notes
+
+# v0.1.0
 This is initial version of the project containing only the basic infrastructure
 setup and functionality.